Added primary term helper to avoid duplicate code.
