+++ src/builders/primary-term-builder.php
@@ -2,6 +2,7 @@
 
 namespace Yoast\WP\SEO\Builders;
 
+use Yoast\WP\SEO\Helpers\Primary_Term_Helper;
 use Yoast\WP\SEO\Repositories\Primary_Term_Repository;
 
 /**
@@ -17,12 +18,21 @@ class Primary_Term_Builder {
 	protected $repository;
 
 	/**
+	 * The primary term helper.
+	 *
+	 * @var Primary_Term_Helper
+	 */
+	private $primary_term;
+
+	/**
 	 * Primary_Term_Builder constructor.
 	 *
-	 * @param Primary_Term_Repository $repository The primary term repository.
+	 * @param Primary_Term_Repository $repository   The primary term repository.
+	 * @param Primary_Term_Helper     $primary_term The primary term helper.
 	 */
-	public function __construct( Primary_Term_Repository $repository ) {
-		$this->repository = $repository;
+	public function __construct( Primary_Term_Repository $repository, Primary_Term_Helper $primary_term ) {
+		$this->repository   = $repository;
+		$this->primary_term = $primary_term;
 	}
 
 
@@ -32,52 +42,9 @@ class Primary_Term_Builder {
 	 * @param int $post_id The post ID.
 	 */
 	public function build( $post_id ) {
-		foreach ( $this->get_primary_term_taxonomies( $post_id ) as $taxonomy ) {
+		foreach ( $this->primary_term->get_primary_term_taxonomies( $post_id ) as $taxonomy ) {
 			$this->save_primary_term( $post_id, $taxonomy->name );
 		}
-	}
-
-	/**
-	 * Returns all the taxonomies for which the primary term selection is enabled.
-	 *
-	 * @param int $post_id Default current post ID.
-	 *
-	 * @return array The taxonomies.
-	 */
-	protected function get_primary_term_taxonomies( $post_id = null ) {
-		if ( $post_id === null ) {
-			$post_id = \get_the_ID();
-		}
-
-		$taxonomies = $this->generate_primary_term_taxonomies( $post_id );
-
-		return $taxonomies;
-	}
-
-	/**
-	 * Generate the primary term taxonomies.
-	 *
-	 * @param int $post_id ID of the post.
-	 *
-	 * @return array The taxonomies.
-	 */
-	protected function generate_primary_term_taxonomies( $post_id ) {
-		$post_type      = \get_post_type( $post_id );
-		$all_taxonomies = \get_object_taxonomies( $post_type, 'objects' );
-		$all_taxonomies = \array_filter( $all_taxonomies, [ $this, 'filter_hierarchical_taxonomies' ] );
-
-		/**
-		 * Filters which taxonomies for which the user can choose the primary term.
-		 *
-		 * @api array    $taxonomies An array of taxonomy objects that are primary_term enabled.
-		 *
-		 * @param string $post_type      The post type for which to filter the taxonomies.
-		 * @param array  $all_taxonomies All taxonomies for this post types, even ones that don't have primary term
-		 *                               enabled.
-		 */
-		$taxonomies = (array) \apply_filters( 'wpseo_primary_term_taxonomies', $all_taxonomies, $post_type, $all_taxonomies );
-
-		return $taxonomies;
 	}
 
 	/**
