+++ src/integrations/watchers/primary-term-watcher.php
@@ -10,6 +10,7 @@ namespace Yoast\WP\SEO\Integrations\Watchers;
 use WPSEO_Meta;
 use Yoast\WP\SEO\Builders\Primary_Term_Builder;
 use Yoast\WP\SEO\Conditionals\Migrations_Conditional;
+use Yoast\WP\SEO\Helpers\Primary_Term_Helper;
 use Yoast\WP\SEO\Helpers\Site_Helper;
 use Yoast\WP\SEO\Integrations\Integration_Interface;
 use Yoast\WP\SEO\Repositories\Primary_Term_Repository;
@@ -34,6 +35,13 @@ class Primary_Term_Watcher implements Integration_Interface {
 	protected $site;
 
 	/**
+	 * The primary term helper.
+	 *
+	 * @var Primary_Term_Helper
+	 */
+	private $primary_term;
+
+	/**
 	 * The primary term builder.
 	 *
 	 * @var Primary_Term_Builder
@@ -54,15 +62,18 @@ class Primary_Term_Watcher implements Integration_Interface {
 	 *
 	 * @param Primary_Term_Repository $repository           The primary term repository.
 	 * @param Site_Helper             $site                 The site helper.
+	 * @param Primary_Term_Helper     $primary_term         The primary term helper.
 	 * @param Primary_Term_Builder    $primary_term_builder The primary term builder.
 	 */
 	public function __construct(
 		Primary_Term_Repository $repository,
 		Site_Helper $site,
+		Primary_Term_Helper $primary_term,
 		Primary_Term_Builder $primary_term_builder
 	) {
 		$this->repository           = $repository;
 		$this->site                 = $site;
+		$this->primary_term         = $primary_term;
 		$this->primary_term_builder = $primary_term_builder;
 	}
 
@@ -82,7 +93,7 @@ class Primary_Term_Watcher implements Integration_Interface {
 	 * @return void
 	 */
 	public function delete_primary_terms( $post_id ) {
-		foreach ( $this->get_primary_term_taxonomies( $post_id ) as $taxonomy ) {
+		foreach ( $this->primary_term->get_primary_term_taxonomies( $post_id ) as $taxonomy ) {
 			$primary_term = $this->repository->find_by_post_id_and_taxonomy( $post_id, $taxonomy->name, false );
 
 			if ( ! $primary_term ) {
@@ -91,49 +102,6 @@ class Primary_Term_Watcher implements Integration_Interface {
 
 			$primary_term->delete();
 		}
-	}
-
-	/**
-	 * Returns all the taxonomies for which the primary term selection is enabled.
-	 *
-	 * @param int $post_id Default current post ID.
-	 *
-	 * @return array The taxonomies.
-	 */
-	protected function get_primary_term_taxonomies( $post_id = null ) {
-		if ( $post_id === null ) {
-			$post_id = \get_the_ID();
-		}
-
-		$taxonomies = $this->generate_primary_term_taxonomies( $post_id );
-
-		return $taxonomies;
-	}
-
-	/**
-	 * Generate the primary term taxonomies.
-	 *
-	 * @param int $post_id ID of the post.
-	 *
-	 * @return array The taxonomies.
-	 */
-	protected function generate_primary_term_taxonomies( $post_id ) {
-		$post_type      = \get_post_type( $post_id );
-		$all_taxonomies = \get_object_taxonomies( $post_type, 'objects' );
-		$all_taxonomies = \array_filter( $all_taxonomies, [ $this, 'filter_hierarchical_taxonomies' ] );
-
-		/**
-		 * Filters which taxonomies for which the user can choose the primary term.
-		 *
-		 * @api array    $taxonomies An array of taxonomy objects that are primary_term enabled.
-		 *
-		 * @param string $post_type      The post type for which to filter the taxonomies.
-		 * @param array  $all_taxonomies All taxonomies for this post types, even ones that don't have primary term
-		 *                               enabled.
-		 */
-		$taxonomies = (array) \apply_filters( 'wpseo_primary_term_taxonomies', $all_taxonomies, $post_type, $all_taxonomies );
-
-		return $taxonomies;
 	}
 
 	/**
